Create first_hello.py
In this code, we can get the using of print() function in the python. The purpose of using python programming language is that python is the interpreted, object-orientated, high level programming language with dynamic sematic. By using print() function anybody can print the normal calculation, can print the single or multiple statement at using the one print() function in the python. In additional, print() function is use to compare. That all about this code.