Create Variable_in_python.py
Variable in the python is the specific name give to anything to print same variable name in all code easily. While starting the variable name we have to focused on some rule like do not start variable name using any number, special character, it will show the error. That all about the variable.